Note where the user may pull BT firmware if not provided by their distro
The Bluetooth adapter has been working on/off for a while when the console is moved back to the miniuart. With an additional patch to always init the pl011, it works more reliably, but the adapter itself isn't reliable unless the latest BT firmware from the rpi foundation is loaded when the adapter is setup.

Note where these files may be found.